Add candidate interface and related types
