+++ libs/api-interfaces/src/lib/types/digitalSkills.ts
@@ -1,0 +1,1 @@
+export type digitalSkills = 'Dårlig' | 'Gjennomsnittelig' | 'Bra';
