+++ libs/api-interfaces/src/lib/types/education.ts
@@ -1,0 +1,1 @@
+export type education = 'Barneskole' | 'Ungdomsskole' | 'Videregående' | 'Høyskole/Universitet' | 'Ingen';
